Refactor users router to usual Express pattern
The pattern I had previously works well for alex-c-line, but it may be a bit much for this case potentially.
