+++ apps/back-end/src/server/routes/index.ts
@@ -1,6 +1,6 @@
 import type { Express } from "express";
 
-import users from "src/server/routes/users";
+import usersRouter from "src/server/routes/users";
 import defineEndpoint from "src/utility/defineEndpoint";
 import handleMiddleware from "src/utility/handleMiddleware";
 
@@ -11,7 +11,7 @@ function createEndpoints(app: Express) {
       response.status(200).send({ hello: "world" });
     }),
   );
-  app.use(defineEndpoint("users"), users());
+  app.use(defineEndpoint("users"), usersRouter);
 }
 
 export default createEndpoints;
